moving testing files into tests folder
